requirements.txt
python-telegram-bot==13.15
pymongo==4.10.1
python-dotenv==1.0.1
requests==2.32.3
apscheduler==3.6.3